# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AcceptanceSamplingRequest": {
        "additionalProperties": false,
        "description": "Judge a plan at the two quality levels it exists to discriminate.",
        "properties": {
          "aql": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aql",
            "type": "number"
          },
          "ltpd": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ltpd",
            "type": "number"
          },
          "model": {
            "default": "binomial",
            "enum": [
              "binomial",
              "hypergeometric",
              "poisson"
            ],
            "title": "Model",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/SamplingPlanIn"
          }
        },
        "required": [
          "plan",
          "aql",
          "ltpd"
        ],
        "title": "AcceptanceSamplingRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LotInspectionRequest": {
        "additionalProperties": false,
        "description": "Apply a plan to one observed sample. The result is a decision.",
        "properties": {
          "defectives": {
            "minimum": 0.0,
            "title": "Defectives",
            "type": "integer"
          },
          "model": {
            "default": "binomial",
            "enum": [
              "binomial",
              "hypergeometric",
              "poisson"
            ],
            "title": "Model",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/SamplingPlanIn"
          }
        },
        "required": [
          "plan",
          "defectives"
        ],
        "title": "LotInspectionRequest",
        "type": "object"
      },
...
      "SamplingPlanDesignRequest": {
        "additionalProperties": false,
        "description": "Two risk points in, the smallest plan meeting both out.",
        "properties": {
          "aql": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aql",
            "type": "number"
          },
          "consumer_risk": {
            "default": 0.1,
            "title": "Consumer Risk",
            "type": "number"
          },
          "lot_size": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Size"
          },
          "ltpd": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ltpd",
            "type": "number"
          },
          "model": {
            "default": "binomial",
            "enum": [
              "binomial",
              "hypergeometric",
              "poisson"
            ],
            "title": "Model",
            "type": "string"
          },
          "producer_risk": {
            "default": 0.05,
            "title": "Producer Risk",
            "type": "number"
          }
        },
        "required": [
          "aql",
          "ltpd"
        ],
        "title": "SamplingPlanDesignRequest",
        "type": "object"
      },
...
      "SamplingPlanOut": {
        "properties": {
          "acceptance_number": {
            "title": "Acceptance Number",
            "type": "integer"
          },
          "lot_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Size"
          },
          "rejection_number": {
            "title": "Rejection Number",
            "type": "integer"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          }
        },
        "required": [
          "sample_size",
          "acceptance_number",
          "lot_size",
          "rejection_number"
        ],
        "title": "SamplingPlanOut",
        "type": "object"
      },
      "SamplingPlanReportOut": {
        "properties": {
          "aoql": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AOQLimitOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "aql": {
            "title": "Aql",
            "type": "number"
          },
          "ati_at_aql": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ati At Aql"
          },
          "consumer_risk": {
            "title": "Consumer Risk",
            "type": "number"
          },
          "indifference_quality": {
            "title": "Indifference Quality",
            "type": "number"
          },
          "ltpd": {
            "title": "Ltpd",
            "type": "number"
          },
          "model": {
            "enum": [
              "binomial",
              "hypergeometric",
              "poisson"
            ],
            "title": "Model",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/SamplingPlanOut"
          },
          "probability_accept_at_aql": {
            "title": "Probability Accept At Aql",
            "type": "number"
          },
          "probability_accept_at_ltpd": {
            "title": "Probability Accept At Ltpd",
            "type": "number"
          },
          "producer_risk": {
            "title": "Producer Risk",
            "type": "number"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "plan",
          "model",
          "aql",
          "ltpd",
          "producer_risk",
          "consumer_risk",
          "probability_accept_at_aql",
          "probability_accept_at_ltpd",
          "indifference_quality",
          "aoql",
          "ati_at_aql",
          "warnings"
        ],
        "title": "SamplingPlanReportOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stateless HTTP surface over capstat-core: reference-validated SPC, process-capability and run-rule statistics, plus CSV/XLSX ingestion. Every response mirrors a core dataclass faithfully, warnings included.",
    "title": "capstat API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/compute/acceptance-sampling/design": {
      "post": {
        "operationId": "compute_acceptance_sampling_design_compute_acceptance_sampling_design_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingPlanDesignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingPlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Acceptance Sampling Design",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/acceptance-sampling/evaluate": {
      "post": {
        "operationId": "compute_acceptance_sampling_evaluate_compute_acceptance_sampling_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptanceSamplingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingPlanReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Acceptance Sampling Evaluate",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/acceptance-sampling/inspect": {
      "post": {
        "operationId": "compute_acceptance_sampling_inspect_compute_acceptance_sampling_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotInspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotDecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Acceptance Sampling Inspect",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/acceptance-sampling/oc-curve": {
      "post": {
        "operationId": "compute_acceptance_sampling_oc_curve_compute_acceptance_sampling_oc_curve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OCCurveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCCurveOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Acceptance Sampling Oc Curve",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/bias": {
      "post": {
        "operationId": "compute_bias_compute_bias_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiasReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Bias",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/capability": {
      "post": {
        "operationId": "compute_capability_compute_capability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Capability",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/capability/analyze": {
      "post": {
        "operationId": "compute_capability_analyze_compute_capability_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeCapabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityAnalysisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Capability Analyze",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/control-chart/cusum": {
      "post": {
        "operationId": "compute_cusum_compute_control_chart_cusum_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CusumRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CusumChartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Cusum",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/control-chart/ewma": {
      "post": {
        "operationId": "compute_ewma_compute_control_chart_ewma_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EwmaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EwmaChartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Ewma",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/control-chart/i-mr": {
      "post": {
        "operationId": "compute_i_mr_compute_control_chart_i_mr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IMRRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartPairOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute I Mr",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/control-chart/xbar-r": {
      "post": {
        "operationId": "compute_xbar_r_compute_control_chart_xbar_r_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubgroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartPairOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Xbar R",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/control-chart/xbar-s": {
      "post": {
        "operationId": "compute_xbar_s_compute_control_chart_xbar_s_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubgroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartPairOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Xbar S",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/descriptive": {
      "post": {
        "operationId": "compute_descriptive_compute_descriptive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptiveSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Descriptive",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/gage-rr": {
      "post": {
        "operationId": "compute_gage_rr_compute_gage_rr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GageRRRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GageRRReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Gage Rr",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/linearity": {
      "post": {
        "operationId": "compute_linearity_compute_linearity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinearityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinearityReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Linearity",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/rules/nelson": {
      "post": {
        "operationId": "compute_nelson_rules_compute_rules_nelson_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleViolationOut"
                  },
                  "title": "Response Compute Nelson Rules Compute Rules Nelson Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Nelson Rules",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/rules/western-electric": {
      "post": {
        "operationId": "compute_western_electric_rules_compute_rules_western_electric_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleViolationOut"
                  },
                  "title": "Response Compute Western Electric Rules Compute Rules Western Electric Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Western Electric Rules",
        "tags": [
          "compute"
        ]
      }
    },
    "/compute/stability": {
      "post": {
        "operationId": "compute_stability_compute_stability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StabilityReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Stability",
        "tags": [
          "compute"
        ]
      }
    },
...
  }
}